+++ Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "manga_updater"
 version = "0.1.0"
-authors = ["Yvan Sanson <yvan.sanson@polytech.univ-montp2.fr>"]
+authors = ["Yvan Sanson <sansonyvan@gmail.com>"]
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -11,3 +11,5 @@ scraper = "0.12.0"
 reqwest = { version = "0.10", features = ["blocking", "json"] }
 tokio = { version = "0.2", features = ["full"] }
 futures = "0.3.7"
+structopt = "0.3"
+csv = "1.1"
