+++ src/file_reader/mod.rs
@@ -1,13 +1,14 @@
 use std::io;
 use std::fs;
+use std::path::PathBuf;
 
 
-pub fn read_urls(file_path: Option<& std::string::String>) -> Result<String, io::Error> {
+pub fn read_urls(file_path: Option<PathBuf>) -> Result<String, io::Error> {
 
-    let mut path: String = String::from("mangas.txt");
+    let mut path: PathBuf = PathBuf::from("mangas.txt");
 
     if file_path.is_some() {
-        path = file_path.unwrap().to_string();
+        path = file_path.unwrap();
     }
     let contents = fs::read_to_string(path)?;
